Fix toc-accordion bugs and move inner-menu-model to separate file
This should fix the remaining problems with the toc-accordion menu:
- The menu sometimes not loading at all when refreshing the page or coming to a page via a direct url.
- An adjacent sub-branch to the branch where the current item lies missing completely after page refresh or coming to a page via a direct url.
