+++ src/components/table-of-contents-accordion/table-of-contents-accordion.html
@@ -55,7 +55,7 @@
                         <span *ngIf="menuItem.description" class="description">{{ menuItem.description }}</span>
                     </ion-item>
                 </ng-template>
-                <ng-template [ngIf]="(menuItem.childrenCount && menuItem.subOptions.length) || menuItem.children_id || menuItem.search_children_id">
+                <ng-template [ngIf]="(menuItem.childrenCount && menuItem.subOptions.length) || menuItem.children_id > -1 || menuItem.search_children_id > -1">
                     <ion-list no-margin *ngIf="menuItem.type!='hidden'" class="accordion-menu" [style.marginLeft]="menuItem.is_child ? 20 + 'px' : '0px'">
                         <ion-item class="header type-{{menuItem.type}}" [ngClass]="menuSettings?.showSelectedOption && menuItem.selected ? menuSettings.selectedOptionClass : null" (tap)="toggleItemOptions(menuItem)" tappable>
                             <ion-icon [class.rotate]="(menuItem.expanded || menuItem.collapsed === false)" class="header-icon" [name]="'ios-arrow-forward'" small item-right></ion-icon>
