+++ src/components/table-of-contents-accordion/table-of-contents-accordion.ts
@@ -6,179 +6,13 @@ import { SideMenuRedirectEvent, SideMenuRedirectEventData } from '../../app/mode
 import { Storage } from '@ionic/storage';
 import { GeneralTocItem } from '../../app/models/table-of-contents.model';
 import { TocAccordionMenuOptionModel } from '../../app/models/toc-accordion-menu-option.model';
+import { InnerMenuOptionModel } from '../../app/models/inner-menu-option.model';
 import { ConfigService,  } from '@ngx-config/core';
 import { LanguageService } from '../../app/services/languages/language.service';
 import { UserSettingsService } from '../../app/services/settings/user-settings.service';
 import { ThrowStmt } from '@angular/compiler';
 import { TranslateModule, TranslatePipe, TranslateService } from '@ngx-translate/core';
 import { MetadataService } from '../../app/services/metadata/metadata.service';
-
-/*-------------------------------------*/
-
-// All children will be stored here in order to reduce lag
-// They will be used everytime a parent is toggled in toggleItemOptions
-const childrenToc = {};
-const searchChildrenToc = {};
-let childrenTocIdCounter = 1;
-
-class InnerMenuOptionModel {
-  public static counter = 1;
-
-  id?: number;
-  iconName?: string;
-  text?: string;
-  title?: string;
-  targetOption?: TocAccordionMenuOptionModel;
-  parent?: InnerMenuOptionModel;
-  selected?: boolean;
-  expanded?: boolean;
-  childrenCount?: number;
-  subOptions?: Array<InnerMenuOptionModel>;
-  type?: string;
-  publication_id?: any;
-  facsimile_id?: any;
-  page_nr?: any;
-  facs_nr?: any;
-  children?: Array<InnerMenuOptionModel>;
-  is_child?: boolean;
-  is_gallery?: boolean;
-  itemId?: any;
-  markdownID?: any;
-  datafile?: any;
-  url?: any;
-  song_id?: any;
-  amountOfParents?: any;
-  openByDefault?: boolean;
-  loading?: boolean;
-  children_id?: any;
-  search_children_id?: any;
-  important?: boolean;
-  description: any;
-  collapsed: boolean;
-
-  public static fromMenuOptionModel(
-                  option: TocAccordionMenuOptionModel,
-                  parent?: InnerMenuOptionModel,
-                  isChild?: boolean,
-                  searchingTocItem?: Boolean
-                ): InnerMenuOptionModel {
-
-    const innerMenuOptionModel = new InnerMenuOptionModel();
-
-    innerMenuOptionModel.id = this.counter++;
-    innerMenuOptionModel.targetOption = option;
-    innerMenuOptionModel.parent = parent || null;
-    innerMenuOptionModel.type = option.type;
-    innerMenuOptionModel.selected = option.selected || false;
-
-    innerMenuOptionModel.important = option.important ? option.important : false;
-
-    if (['file', 'folder'].indexOf(option.type) !== -1) {
-      innerMenuOptionModel.markdownID = option.id;
-    }
-
-    if (option.text) {
-      innerMenuOptionModel.text = option.text;
-      innerMenuOptionModel.description = option.description;
-    } else if (option.title) {
-      innerMenuOptionModel.text = option.title;
-    }
-
-    innerMenuOptionModel.publication_id = option.publication_id || null;
-    innerMenuOptionModel.page_nr = option.page_nr || null;
-    innerMenuOptionModel.facs_nr = option.facs_nr || null;
-
-    innerMenuOptionModel.is_gallery = option.is_gallery || false;
-
-    if (option.itemId) {
-      innerMenuOptionModel.itemId = option.itemId;
-
-      const legacyID = option.itemId;
-      const legacyData = legacyID.split('_');
-
-      if (legacyData.length === 2 && innerMenuOptionModel.publication_id === null) {
-        innerMenuOptionModel.publication_id = legacyData[1];
-      }
-    }
-
-    if (option.facsimile_id) {
-      innerMenuOptionModel.facsimile_id = option.facsimile_id;
-    }
-
-    if (option.song_id) {
-      innerMenuOptionModel.song_id = option.song_id;
-    }
-
-    if (option.openByDefault) {
-      innerMenuOptionModel.openByDefault = true;
-    }
-
-    if (isChild) {
-      innerMenuOptionModel.is_child = true;
-    } else {
-      innerMenuOptionModel.is_child = false;
-    }
-
-    if (option.collapsed === false) {
-      innerMenuOptionModel.collapsed = false;
-      innerMenuOptionModel.expanded = true;
-      innerMenuOptionModel.important = true; // SK 18.5.2022 added
-    } else {
-      innerMenuOptionModel.collapsed = true;
-      innerMenuOptionModel.expanded = false;
-    }
-
-    if (option.children) {
-      // innerMenuOptionModel.expanded = false;
-      let storeChildren = false;
-
-      if (option.openByDefault) {
-      // innerMenuOptionModel.expanded = true;
-      }
-
-      innerMenuOptionModel.childrenCount = option.children.length;
-      innerMenuOptionModel.subOptions = [];
-
-      if (
-        option.children && option.children.length &&
-        !option.children_id && !option.search_children_id &&
-        !innerMenuOptionModel.markdownID &&
-        !searchingTocItem
-      ) {
-        innerMenuOptionModel.children_id = childrenTocIdCounter;
-        childrenToc[innerMenuOptionModel.children_id] = [];
-        childrenTocIdCounter++;
-        storeChildren = true;
-      }
-
-      option.children.forEach(subItem => {
-        const innerSubItem = InnerMenuOptionModel.fromMenuOptionModel(subItem, innerMenuOptionModel, true, searchingTocItem);
-        // innerSubItem.collapsed = subItem.collapsed;
-        if (innerSubItem.text) {
-          if (storeChildren) {
-            if ( childrenToc[innerMenuOptionModel.children_id].indexOf(innerSubItem) === -1 ) {
-              childrenToc[innerMenuOptionModel.children_id].push(innerSubItem);
-            }
-          } else {
-            if ( innerMenuOptionModel.subOptions.indexOf(innerSubItem) === -1 ) {
-              innerMenuOptionModel.subOptions.push(innerSubItem);
-            }
-          }
-        }
-
-        // Expand the parent if any
-        // child option is selected
-        if (subItem.selected) {
-          // innerSubItem.parent.selected = true;
-          innerSubItem.parent.expanded = true;
-          innerSubItem.parent.collapsed = false;
-        }
-      });
-    }
-    return innerMenuOptionModel;
-  }
-}
-
 
 @Component({
   selector: 'table-of-contents-accordion',
@@ -213,6 +47,7 @@ export class TableOfContentsAccordionComponent {
       // Map the options to our internal models
       this.menuOptions.forEach(option => {
         const innerMenuOption = InnerMenuOptionModel.fromMenuOptionModel(option, null, false, value.searchTocItem);
+        // console.log(innerMenuOption);
         if ( this.collapsableItems.indexOf(innerMenuOption) === -1 ) {
           this.collapsableItems.push(innerMenuOption);
         }
@@ -274,6 +109,7 @@ export class TableOfContentsAccordionComponent {
               this.collapsableItems[i] = innerMenuOptionWithoutChildren;
             }
           }
+          // console.log('collapsable items', this.collapsableItems);
         }
 
       }
@@ -298,6 +134,12 @@ export class TableOfContentsAccordionComponent {
   @Input() isGallery?: Boolean;
   @Input() open: Boolean;
   @Output() selectOption = new EventEmitter<any>();
+
+  // All children will be stored here in order to reduce lag
+  // They will be used everytime a parent is toggled in toggleItemOptions
+  childrenToc = {};
+  searchChildrenToc = {};
+  // childrenTocIdCounter = 1;
 
   currentItem: GeneralTocItem;
   currentOption: any;
@@ -448,7 +290,7 @@ export class TableOfContentsAccordionComponent {
   }
 
   ngOnChanges(about) {
-    if ( Array.isArray(about) ) {
+    if ( Array.isArray(about) && this.isMarkdown) {
       // console.log('toc-accordion ngOnChanges initialized', about);
       this.menuOptions = about;
       this.collapsableItems = new Array<InnerMenuOptionModel>();
@@ -740,7 +582,7 @@ export class TableOfContentsAccordionComponent {
         for (const child of tocItem.parent.subOptions) {
           if (child.subOptions && child.subOptions.length && !child.important) {
             child['search_children_id'] = this.tocItemSearchChildrenCounter;
-            searchChildrenToc[child.search_children_id] = child.subOptions;
+            this.searchChildrenToc[child.search_children_id] = child.subOptions;
             child.subOptions = [];
             this.tocItemSearchChildrenCounter++;
           }
@@ -1073,11 +915,11 @@ export class TableOfContentsAccordionComponent {
       } else if (option.childrenCount) {
         if (option.subOptions.length) {
           this.selectOneItem(itemId, option.subOptions, option.subOptions);
-        } else if (!option.subOptions.length) {
-          if (option.children_id) {
-            option.subOptions = childrenToc[option.children_id];
-          } else if (option.search_children_id) {
-            option.subOptions = searchChildrenToc[option.search_children_id];
+        } else {
+          if (option.search_children_id > -1) {
+            option.subOptions = this.searchChildrenToc[option.search_children_id];
+          } else if (option.children_id > -1) {
+            option.subOptions = this.childrenToc[option.children_id];
           }
           this.selectOneItem(itemId, option.subOptions, option.subOptions);
         }
@@ -1114,10 +956,10 @@ export class TableOfContentsAccordionComponent {
 
     // Fetch suboptions if item doesn't have them already
     if (!targetOption.subOptions.length) {
-      if (targetOption.children_id) {
-        targetOption.subOptions = childrenToc[targetOption.children_id];
-      } else if (targetOption.search_children_id) {
-        targetOption.subOptions = searchChildrenToc[targetOption.search_children_id];
+      if (targetOption.search_children_id > -1) {
+        targetOption.subOptions = this.searchChildrenToc[targetOption.search_children_id];
+      } else if (targetOption.children_id > -1) {
+        targetOption.subOptions = this.childrenToc[targetOption.children_id];
       }
     }
     if ( targetOption.collapsed === undefined || String(targetOption.collapsed) === '' ) {
@@ -1213,12 +1055,12 @@ export class TableOfContentsAccordionComponent {
       this.selectedOption.selected = false;
       this.selectedOption.targetOption.selected = false;
 
-    if (this.selectedOption.parent) {
-      this.selectedOption.parent.selected = false;
-      this.selectedOption.parent.expanded = false;
-    }
+      if (this.selectedOption.parent) {
+        this.selectedOption.parent.selected = false;
+        this.selectedOption.parent.expanded = false;
+      }
 
-    this.selectedOption = null;
+      this.selectedOption = null;
     }
 
     // Set this option to be the selected
